requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alpaca-py==0.43.2
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
idna==3.11
    # via requests
msgpack==1.1.2
    # via alpaca-py
numpy==2.4.1
    # via pandas
pandas==2.3.3
    # via
    #   -r requirements.in
    #   alpaca-py
pydantic==2.12.5
    # via alpaca-py
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
requests==2.32.5
    # via alpaca-py
six==1.17.0
    # via python-dateutil
sseclient-py==1.9.0
    # via alpaca-py
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
websockets==16.0
    # via alpaca-py
